create util for prompting user to select a project
