+++ lib/init/project.js
@@ -2,7 +2,6 @@ const fs = require('fs');
 
 const ora = require('ora');
 const { YAMLException } = require('js-yaml');
-const { AutoComplete } = require('enquirer');
 
 const api = require('../api').default;
 const config = require('../config');
@@ -10,6 +9,7 @@ const consts = require('../consts');
 const output = require('../output');
 const { collectAndSaveToken } = require('../init/token');
 const getSelectedProjects = require('../utils/getSelectedProjects');
+const promptForProject = require('../utils/promptForProject');
 
 function quit() {
   console.log('\nExiting Ditto CLI...');
@@ -80,11 +80,6 @@ async function getProjects(token, projectsAlreadySelected) {
   return projectsUnselected.map((item) => `${item.name} ${output.subtle(item.url)}`);
 }
 
-function parseResponse(response) {
-  const [, name, id] = response.split(/^(.*)\s.*http.*\/(\w+).*$/);
-  return { name, id };
-}
-
 async function collectProject(token, initialize) {
   const path = process.cwd();
   if (initialize) {
@@ -103,15 +98,10 @@ async function collectProject(token, initialize) {
     return null;
   }
 
-  const prompt = new AutoComplete({
-    name: 'project',
-    message: initialize ? "Choose the project you'd like to sync text from" : 'Add a project',
-    limit: 10,
-    choices: projects,
+  return promptForProject({ 
+    projects,
+    message: initialize ? "Choose the project you'd like to sync text from" : 'Add a project'
   });
-
-  const response = await prompt.run();
-  return parseResponse(response);
 }
 
 async function collectAndSaveProject(initialize = false) {
